Added a new Configuration class. Moved the GetSettings method to this class.
