+++ Endzone.Umbraco.Extensions/PublishedContentExtensions/Collections.cs
@@ -104,18 +104,5 @@ namespace Endzone.Umbraco.Extensions.PublishedContentExtensions
                          select part.AsEnumerable();
             return splits;
         }
-
-        /// <summary>
-        /// Returns the website settings node
-        /// </summary>
-        /// <param name="content"></param>
-        /// <returns></returns>
-        public static IPublishedContent GetWebsiteSettings(this IPublishedContent content)
-        {
-            var umbracoHelper = new UmbracoHelper(UmbracoContext.Current);
-            var homepage = content.AncestorOrSelf(1);
-            var link = homepage.GetLink("websiteSettings");
-            return umbracoHelper.TypedContent(link.Id);
-        }
     }
 }
